+++ CHANGELOG.md
@@ -16,6 +16,8 @@
 - Bot Combat Tech health and speed bonuses not being applied to the Gil-E
 - Complete Bot Tech tooltip not listing the Stinger or Gil-E
 - Complete Orbital Tech tooltip not listing the ARKYD
+- Stinger Upgrade tooltip
+- Correct tooltips for Vehicle Factory and Advanced Vehicle Factory being reversed
 
 ## v5.26.1 - 2021-12-08